Time: 496 ms (92.42%) | Memory: 0B (100.00%) - LeetSync

+++ 1182-game-play-analysis-iv/game-play-analysis-iv.sql
@@ -19,5 +19,5 @@ WITH first_logins AS (
 SELECT
   ROUND(
     ((SELECT C.num_logins FROM consec_logins C)
-    / (SELECT COUNT(F.player_id) FROM first_logins F)::FLOAT)::NUMERIC
+    / (SELECT COUNT(F.player_id) FROM first_logins F)::NUMERIC)
   , 2) AS fraction
